pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "natspy-lock"
dynamic = ["version"]
description = 'Distributed lock library using nats.py'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = [
  "Distributed lock",
  "nats.py",
  "python lock"
]
authors = [
  { name = "Seungwoo Hong", email = "qksn1541@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "nats-py>=2.0.0"
]

[project.urls]
Documentation = "https://github.com/HHongSeungWoo/natspy-lock#readme"
Issues = "https://github.com/HHongSeungWoo/natspy-lock/issues"
Source = "https://github.com/HHongSeungWoo/natspy-lock"

[tool.hatch.version]
path = "natspy_lock/__init__.py"

[tool.hatch.envs.default]
path = "venv"
type = "virtual"
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "pytest-asyncio",
  "pre-commit<4"
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.7", "3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.lint]
path = "venv"
type = "virtual"
detached = true
dependencies = [
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:natspy_lock tests}"
style = [
  "ruff {args:.}",
]
fmt = [
  "ruff --fix {args:.}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.ruff]
target-version = "py37"
line-length = 120
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["natspy_lock"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
"__init__.py" = ["F401"]

[tool.coverage.run]
source_pkgs = ["natspy_lock", "tests"]
branch = true
parallel = true
omit = []

[tool.coverage.paths]
natspy_lock = ["natspy_lock", "*/natspy-lock/natspy_lock"]
tests = ["tests", "natspy-lock/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-s"
